fix sieve

+++ src/problem051.h
@@ -21,7 +21,7 @@
  */
 
 namespace problem051 {
-    int getAnswer();
+    long getAnswer();
 };
